// src/interfaces/survey-answer.interface.ts
export interface SurveyAnswer {
  survey_log_id: string;
  survey_question_id: string;
  unique_member_id: string;
  ik_number: string;
  answer: string;
  date_logged: string;
  staff_id: string;
  app_version: string;
  operator_id: string;
  desc: string;
  unique_entity_id: string;
  entity_id: string;
  staff_entity_id: string;
  section: string;
}